[bug fix] fix regression in findHarmonics
Changing the datatype of the locations in findPeaks to uint16 caused a
datatype saturation in findHarmonics because the subtraction in line
19 returned 0 instead of a negative number. Changing to int32 fixes the
regression.

+++ findPeaks.m
@@ -43,7 +43,7 @@ function [locations] = localMaxima1d(x)
 %   - locations: the indices of each maxima/peak
 
 % preallocate; there can't be more maxima than half the size of x
-locations = zeros(1, floor(numel(x)/2), 'uint16');
+locations = zeros(1, floor(numel(x)/2), 'int32');
 nPeaksIdx = 1;
